Moving onRest to listener-misc-handler
